Mouse DoRothEA

A table reporting signed human TF-target interactions. This database covers in total 1179 TFs targeting 17,410 genes with 410,455 unique interactions. In addition, each TF is accompanied with an emperical confidence level that was derived from the number of supporting evidences for this TF/interaction. The range is from A (high quality) to E (low quality).

Format

A table of mouse TF-target interactions:

tf

TF identifier as MGI symbols

confidence

summary confidence score classifying regulons based on their quality

target

target identifier as MGI symbols

mor

mode of regulation indicating the effect of a TF on the target

Source

https://www.ncbi.nlm.nih.gov/pubmed/31525460